better structure for token docs files
